# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is Api for Autogen.",
        "title": "Autogen API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/user/amr": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "FindByAMRID All AMR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AMR"
                ],
                "summary": "FindByAMRID All AMR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search Parameter",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting Parameter",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Rows Count",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_ListAMRResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Upload AMR dataset",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AMR"
                ],
                "summary": "Upload AMR dataset",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File Excel AMR (.xlsx/.xls/.csv)",
                        "name": "attachment",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_UploadAMRResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_UploadAMRResp"
                        }
                    }
                }
            }
        },
        "/user/amr/:amr_id": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Find AMR",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AMR"
                ],
                "summary": "Find AMR",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "AMR ID",
                        "name": "amr_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_FindAMRResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_FindAMRResp"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete AMR",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AMR"
                ],
                "summary": "Delete AMR",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "AMR ID",
                        "name": "amr_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_DeleteAMRResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_DeleteAMRResp"
                        }
                    }
                }
            }
        },
        "/user/amr/:amr_id/generate-report": {
            "post": {
                "description": "Create Generate Report AMR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AMR"
                ],
                "summary": "Generate Report AMR",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "AMR ID",
                        "name": "amr_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Generare Report of AMR Request Parameter",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logisfy_internal_model_request.GenerateReportAMRReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_GenerateReportAMRResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_GenerateReportAMRResp"
                        }
                    }
                }
            }
        },
...
        "/user/amr/:amr_id/report": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Report AMR by AMR ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AMR"
                ],
                "summary": "Get Report AMR",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "AMR ID",
                        "name": "amr_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search Parameter",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting Parameter",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Rows Count",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_ListReportAMRResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "logisfy_core_response.ApiResponse-logisfy_internal_model_response_ListAMRResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logisfy_internal_model_response.ListAMRResp"
                },
                "error": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "tin": {
                    "type": "string"
                },
                "tout": {
                    "type": "string"
                }
            }
        },
...
        "logisfy_core_response.ApiResponse-logisfy_internal_model_response_UploadAMRResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logisfy_internal_model_response.UploadAMRResp"
                },
                "error": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "tin": {
                    "type": "string"
                },
                "tout": {
                    "type": "string"
                }
            }
        },
...
        "logisfy_internal_model_request.GenerateReportAMRReq": {
            "type": "object"
        },
...
        "logisfy_internal_model_response.UploadAMRResp": {
            "type": "object",
            "properties": {
                "amr_id": {
                    "type": "integer"
                },
                "data_count": {
                    "type": "integer"
                }
            }
        },
...
    }
}